services/analyzer/pyproject.toml
[project]
name = "video-verification-analyzer"
version = "1.0.0"
description = "Analyzer service: frames + images → VLM analysis"
requires-python = ">=3.11"
dependencies = [
    "video-verification-shared",
    "confluent-kafka>=2.3",
    "httpx>=0.27",
    "Pillow>=10.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["analyzer*"]
